Public-portfolio/Finance/Fund_Manager_Monitoring/manager_monitoring/requirements.txt
# Core
pandas>=2.0
numpy>=1.24
matplotlib>=3.7

# Reporting
python-docx>=1.1

# Performance analytics
pyfolio-reloaded>=0.9


ipykernel>=6.0
